requirements.txt
amqp==2.3.2
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
coveralls==1.5.0
cryptography==2.3.1
cssutils==1.0.2
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11
django-bootstrap4==0.0.7
django-classy-tags==0.8.0
django-filter==2.0.0
django-fontawesome==1.0
django-inlinecss==0.1.2
django-redis-cache==1.7.1
django-sekizai==0.10.0
django-sslserver==0.20
django-tables2==2.0.1
djangorestframework==3.9.0
docopt==0.6.2
eventbrite==3.3.3
factory-boy==2.11.1
Faker==0.9.1
gunicorn==19.7.1
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
kombu==4.2.1
mock==2.0.0
mongoengine==0.15.3
oauthlib==2.1.0
parso==0.3.1
pbr==5.1.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.3.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.7.1
pynliner3==0.6
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
redis==2.10.6
reportlab==3.5.9
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==1.7.0
social-auth-storage-mongoengine==1.0.1
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
wcwidth==0.1.7
whitenoise==3.3.1
